requirements.txt
# Payment app (FastAPI) – minimal dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
SQLAlchemy==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
email-validator==2.0.0
aiosmtplib==3.0.1
boto3==1.34.0
python-multipart==0.0.6
